pyproject.toml
[tool.poetry]
name = "hyperhire-chat"
version = "0.1.0"
description = ""
authors = ["David Onuh <davidonuh1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-filter = "^23.4"
drf-yasg = "^1.21.7"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"
pytest-django = "^4.7.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
